+++ tests/conftest.py
@@ -4,8 +4,8 @@ import pytest
 import random
 import torch
 from gpytorch.mlls import ExactMarginalLogLikelihood
-from creative_project._initializers import Initializers
-from creative_project._validators import Validators
+from greattunes._initializers import Initializers
+from greattunes._validators import Validators
 
 ### Parsing of keywords: allow for specialized tests for different python versions
 def pytest_addoption(parser):
@@ -98,7 +98,7 @@ def training_data_covar_complex(covars_initialization_data):
     """
     defines simple training data that corresponds to covar_complex (covars_initialization_data[1]), where covar_complex
     is the right format for initialization of the full user-facing class CreativeProject
-    (creative_project.CreativeProject)
+    (greattunes.CreativeProject)
     """
 
     covars = covars_initialization_data[1]
@@ -174,7 +174,7 @@ def ref_model_and_multivariate_training_data(training_data_covar_complex):
 @pytest.fixture(scope="module")
 def tmp_observe_class():
     """
-    temporary class to allow testing of methods from creative_project._observe
+    temporary class to allow testing of methods from greattunes._observe
     """
 
     # define class
@@ -185,7 +185,7 @@ def tmp_observe_class():
             self.dtype = torch.double
             self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
-        from creative_project._observe import _get_and_verify_response_input, _get_response_function_input, \
+        from greattunes._observe import _get_and_verify_response_input, _get_response_function_input, \
             _read_response_manual_input, _print_candidate_to_prompt, _read_covars_manual_input, \
             _get_and_verify_covars_input, _get_covars_datapoint, _get_response_datapoint
 
@@ -197,7 +197,7 @@ def tmp_observe_class():
 @pytest.fixture(scope="module")
 def tmp_modeling_class():
     """
-    temporary class to allow testing of methods from creative_project._modeling
+    temporary class to allow testing of methods from greattunes._modeling
     """
 
     class TmpClass:
@@ -215,7 +215,7 @@ def tmp_modeling_class():
                           }
 
         # import method
-        from creative_project._modeling import _set_GP_model
+        from greattunes._modeling import _set_GP_model
 
     # initialize class
     cls = TmpClass()
@@ -241,7 +241,7 @@ def tmp_best_response_class():
             self.best_response = None
 
         # import methods
-        from creative_project._best_response import _find_max_response_value, _update_max_response_value, \
+        from greattunes._best_response import _find_max_response_value, _update_max_response_value, \
             current_best, _update_proposed_data
 
     cls = TmpClass()
@@ -257,7 +257,7 @@ def tmp_Initializers_with_find_max_response_value_class():
     """
 
     class TmpClass(Initializers):
-        from creative_project._best_response import _find_max_response_value
+        from greattunes._best_response import _find_max_response_value
 
     cls = TmpClass()