+++ tests/integration/test_acq_func_functional.py
@@ -1,8 +1,8 @@
 import pytest
 import torch
 from botorch.acquisition import ExpectedImprovement
-from creative_project._acq_func import AcqFunction
-import creative_project._acq_func
+from greattunes._acq_func import AcqFunction
+import greattunes._acq_func
 
 
 def test_AcqFunction__initialize_acq_func_functional(custom_models_simple_training_data_4elements, monkeypatch):
@@ -24,7 +24,7 @@ def test_AcqFunction__initialize_acq_func_functional(custom_models_simple_traini
         self.it_ran = True
         return True
     monkeypatch.setattr(
-        creative_project._acq_func.AcqFunction, "set_acq_func", mock_set_acq_func
+        greattunes._acq_func.AcqFunction, "set_acq_func", mock_set_acq_func
     )
 
     # test that it passes with attribute train_Y being not None (should return True)